Refactored to allow for multiple players
Refactored to allow for multiple players. Player, Rooms (this is odd as
they had two Room lists, so I’ve given each a distinct name) and
Characters have been moved into gameSnapshot. An array of gameSnapshots
now exists in MIRCH, with two new functions - one to access the current
game snapshot (getCurrentGameSnapshot()) and one to move to the next
game snapshot in the list. Two game snapshots are created, although i
can not guarantee that these are identical (we can’t copy across the
objects due to Java’s lack of functional deep copy) so this has to be
managed in the game generation class. Currently only the 1st game
snapshot is played - the second does nothing, its up to peter to sort
out transitions from one to the other.